requirements.txt
Appium-Python-Client==0.36
atomicwrites==1.2.1
attrs==18.2.0
funcsigs==1.0.2
more-itertools==5.0.0
pathlib2==2.3.3
pluggy==0.8.1
py==1.7.0
pytest==4.1.1
scandir==1.9.0
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
